Update an install.sh script creation

+++ api/routes/products/delete-product.js
@@ -2,6 +2,7 @@ var router = require('express').Router();
 const server = require('../../server.js')
 var jsonParser = server.parser.json()
 const route_name = "/products/delete-product"
+const permissions_manager = require("../../utils/permissions-manager.js")
 server.logger(" [INFO] /api" + route_name + " route loaded !")
 
 router.delete('', jsonParser, function (req, res) {
@@ -21,12 +22,21 @@ router.delete('', jsonParser, function (req, res) {
       return res.json({'error': true, 'code': 404})
     } else {
       if (result[0].token === req.query.token) {
-        var sql = `DELETE FROM products WHERE id='${req.body.id}'`;
-        server.con.query(sql, function (err, result) {
-            if (err) {server.logger(" [ERROR] Database error\n  " + err), error = true, response = "Database error"};
-        });
-        server.logger(" [DEBUG] Product " + req.body.id + " deleted from " + IP + " !")
-        return res.json({"error": error, "response": response});
+        permissions_manager.has_permission(req.query.uuid, "DELETEPRODUCT").then(function(result) {
+          if (result) {
+            var sql = `DELETE FROM products WHERE id='${req.body.id}'`;
+            server.con.query(sql, function (err, result) {
+                if (err) {server.logger(" [ERROR] Database error\n  " + err), error = true, response = "Database error"};
+            });
+            server.logger(" [DEBUG] Product " + req.body.id + " deleted from " + IP + " !")
+            return res.json({"error": error, "response": response});
+          } else {
+            return res.json({
+              "error": true,
+              "code": 403
+            })
+          }
+        })
       } else {
         return res.json({'error': true, 'code': 403})
       }
